test(nlpgo): extract shared subprocess e2e harness
Pulls ensureNlpgoBinary + spawn + waitForHealth + SSE-frame parsing
out of eval-trace-id-roundtrip into _nlpgoSubprocess.ts so the B1/B2
prompt-playground e2e (and any future nlpgo wire-level proof) imports
{ startNlpgoSubprocess, collectSSE, hasGo } instead of hand-copying
the ~150-line boot block a third time (PR #4048 already duplicated it
twice).

eval-trace-id-roundtrip refactored onto the helper; re-ran locally
end-to-end, still green. Underscore prefix + no .test.ts suffix so
vitest doesn't pick the helper up as a suite.
